requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-allauth-ui==0.2.2
django-phonenumber-field==7.3.0
django-widget-tweaks==1.5.0
idna==3.6
oauthlib==3.2.2
phonenumberslite==8.13.31
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
